DRYing code and updated player to have per li audio element
Above was intended to reduce the amount of times we need to buffer
Before we were loading the tracks each time they were started, now we only load them once.  Also, upon adding them to the queue we download the metadata.  This also allows for future enhancements involving the playlist.